refactor(wal): optimize WriteBatch and simplify rotation logic - Remove redundant rotation checks and consolidate benchmark code - Improve WriteBatch performance by eliminating intermediate flush steps - Move benchmarks to separate file and remove duplication - Update Write to delegate to WriteBatch for consistency
