+++ wal_test.go
@@ -4,12 +4,10 @@ import (
 	"cmp"
 	"fmt"
 	"maps"
-	"math/rand"
 	"os"
 	"slices"
 	"strconv"
 	"testing"
-	"time"
 
 	"github.com/stretchr/testify/require"
 )
@@ -539,142 +537,3 @@ func TestWriteBatch_RotateWhenTmpIndexAlreadyFull(t *testing.T) {
 		require.True(t, ok)
 	}
 }
-
-func TestWrite_NotRotated(t *testing.T) {
-	dir := t.TempDir()
-	segmentThreshold := 3
-
-	wal, err := NewWAL(Config{
-		Dir:              dir,
-		Prefix:           "log_",
-		SegmentThreshold: segmentThreshold,
-		MaxSegments:      10,
-		IsInSyncDiskMode: false,
-	})
-	require.NoError(t, err)
-
-	// write 1, 2, 3
-	for i := 1; i <= segmentThreshold; i++ {
-		require.NoError(t, wal.Write(uint64(i), fmt.Sprintf("k%d", i), []byte("v")))
-	}
-
-	// tmpIndex len is 3, rotation will be triggered on 4
-	require.Equal(t, 3, len(wal.tmpIndex))
-
-	require.NoError(t, wal.Write(uint64(segmentThreshold+1), "k4", []byte("v")))
-
-	// tmpIndex now is 1, rotate was triggered on 4 write, not on 3
-	require.Equal(t, 1, len(wal.tmpIndex))
-
-	require.NoError(t, wal.Close())
-	dir = t.TempDir()
-	wal, err = NewWAL(Config{
-		Dir:              dir,
-		Prefix:           "log2_",
-		SegmentThreshold: segmentThreshold,
-		MaxSegments:      10,
-		IsInSyncDiskMode: false,
-	})
-	require.NoError(t, err)
-
-	// write 1, 2, 3
-	for i := 1; i <= segmentThreshold; i++ {
-		require.NoError(t, wal.Write(uint64(i), fmt.Sprintf("k%d", i), []byte("v")))
-	}
-
-	batch := []Record{
-		{Index: 4, Key: "k4", Value: []byte("v")},
-		{Index: 5, Key: "k5", Value: []byte("v")},
-	}
-	require.NoError(t, wal.WriteBatch(batch))
-
-	// rotation was triggered before 4 write, so tmpIndex is 2 now
-	require.Equal(t, 2, len(wal.tmpIndex))
-
-	require.NoError(t, wal.WriteBatch([]Record{{6, "k6", []byte("v")}}))
-
-	// tmpIndex now is 0, cause rotation in batch operation also triggered after write
-	// rotation logic on batch write:
-	// 1) check how many records can be written to current segment
-	// 2) write them, then do rotation
-	// 3) if there are some records left, then go to 1)
-	//
-	// This does not break the existing write logic, it just makes the rotation check in the Write method after WriteBatch redundant.
-	// It needed for backward compatibility
-	require.Equal(t, 0, len(wal.tmpIndex))
-}
-
-// getRandomData returns a random byte slice of the given size.
-func getRandomData(sizeBytes int) []byte {
-	if sizeBytes <= 0 {
-		return nil
-	}
-	letters := []byte("abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ")
-	b := make([]byte, sizeBytes)
-	for i := range b {
-		b[i] = letters[rand.Intn(len(letters))]
-	}
-	return b
-}
-
-func benchmarkWALWrite(b *testing.B, batchSize int) {
-	dir := b.TempDir()
-
-	log, err := NewWAL(Config{
-		Dir:              dir,
-		Prefix:           "log_",
-		SegmentThreshold: 10000,
-		MaxSegments:      10000000,
-		IsInSyncDiskMode: false,
-	})
-	require.NoError(b, err)
-	defer func() {
-		require.NoError(b, log.Close())
-	}()
-
-	valueData := getRandomData(1024)
-
-	records := make([]Record, b.N*batchSize)
-	for i := range records {
-		records[i] = Record{
-			Index: uint64(i),
-			Value: valueData,
-			Key:   "key" + strconv.Itoa(i),
-		}
-	}
-
-	b.ReportAllocs()
-	b.ResetTimer()
-
-	start := time.Now()
-	total := 0
-
-	for i := 0; i < len(records); i += batchSize {
-		end := min(i+batchSize, len(records))
-
-		if batchSize == 1 {
-			require.NoError(b, log.Write(records[i].Index, records[i].Key, records[i].Value))
-			total++
-		} else {
-			require.NoError(b, log.WriteBatch(records[i:end]))
-			total += end - i
-		}
-	}
-
-	elapsed := time.Since(start)
-	b.ReportMetric(float64(total)/elapsed.Seconds(), "records/s")
-	b.ReportMetric(elapsed.Seconds(), "sec")
-}
-
-func BenchmarkWal_Write(b *testing.B)           { benchmarkWALWrite(b, 1) }
-func BenchmarkWal_WriteBatch10(b *testing.B)    { benchmarkWALWrite(b, 10) }
-func BenchmarkWal_WriteBatch100(b *testing.B)   { benchmarkWALWrite(b, 100) }
-func BenchmarkWal_WriteBatch1000(b *testing.B)  { benchmarkWALWrite(b, 1000) }
-func BenchmarkWal_WriteBatch10000(b *testing.B) { benchmarkWALWrite(b, 10000) }
-
-// Batching significantly improves WAL write throughput (~2–6x, from ~210k to ~1.25M records/s)
-// by amortizing per-operation overhead. Most of the gains are achieved with batch sizes of 100–1000;
-// larger batches (e.g., 10k) provide only marginal improvements while greatly increasing memory usage and allocations.
-//
-// Per-record allocations remain roughly constant (~6–7 allocs, ~1.7–2.0KB),
-// indicating that batching reduces syscall/overhead cost but does not optimize per-record memory behavior.
